feat: added asset scrape debugger
